Completed Table - Dialog Boxes done
All elements have been added to the Periodic Table.
All elements have a dialog box pop up when clicked on.
Got rid of the raw folder - don't need that no more.
Added the JSON file to the assests folder.
Next step is for the app to read the JSON file and display the element
information in the dialog box.